+++ src/ExcelServiceProvider.php
@@ -1,19 +1,21 @@
 <?php
 
+declare(strict_types=1);
+
 namespace PittacusW\Excel;
 
-use Spatie\LaravelPackageTools\Package;
-use Spatie\LaravelPackageTools\PackageServiceProvider;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Support\ServiceProvider;
+use Maatwebsite\Excel\Excel as LaravelExcel;
 
-class ExcelServiceProvider extends PackageServiceProvider
+class ExcelServiceProvider extends ServiceProvider
 {
-    public function configurePackage(Package $package): void
+    public function register(): void
     {
-        /*
-         * This class is a Package Service Provider
-         *
-         * More info: https://github.com/spatie/laravel-package-tools
-         */
-        $package->name('excel');
+        $this->app->singleton(Excel::class, function (Application $app): Excel {
+            return new Excel($app->make(LaravelExcel::class));
+        });
+
+        $this->app->alias(Excel::class, 'pittacusw.excel');
     }
 }