+++ tests/TestCase.php
@@ -1,37 +1,23 @@
 <?php
 
+declare(strict_types=1);
+
 namespace PittacusW\Excel\Tests;
 
-use Illuminate\Database\Eloquent\Factories\Factory;
+use Maatwebsite\Excel\ExcelServiceProvider as LaravelExcelServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 use PittacusW\Excel\ExcelServiceProvider;
 
 class TestCase extends Orchestra
 {
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'PittacusW\\Excel\\Database\\Factories\\'.class_basename($modelName).'Factory'
-        );
-    }
-
-    protected function getPackageProviders($app)
+    /**
+     * @return array<int, class-string>
+     */
+    protected function getPackageProviders($app): array
     {
         return [
+            LaravelExcelServiceProvider::class,
             ExcelServiceProvider::class,
         ];
-    }
-
-    public function getEnvironmentSetUp($app)
-    {
-        config()->set('database.default', 'testing');
-
-        /*
-         foreach (\Illuminate\Support\Facades\File::allFiles(__DIR__ . '/database/migrations') as $migration) {
-            (include $migration->getRealPath())->up();
-         }
-         */
     }
 }